feat/FRO-45-Conversation: Added a new recorder library
